Refactor /molecules
