+++ frontend/app/components/molecules/RecurringMeeting.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import React, { useState } from 'react';
-import RadioGroup from '../../atoms/RadioGroup';
-import Checkbox from '../../atoms/checkbox';
-import SpinnerInput from '../../atoms/SpinnerInput';
-import styles from "../../../../styles/components/organisms/RecurringMeeting.module.scss";
+import RadioGroup from '../atoms/RadioGroup';
+import Checkbox from '../atoms/checkbox';
+import SpinnerInput from '../atoms/SpinnerInput';
+import styles from "../../../styles/components/organisms/RecurringMeeting.module.scss";
 
 const RecurringMeetingForm: React.FC = () => {
   const [isRecurring, setIsRecurring] = useState(false);
